Improve error handling (when results section not found)
Make sure an error is raised if we cannot find out where to put the results,
that is if we cannot find either a results section or the comment markers.
Also, make the 'resultsSection' optional in the config file, since there's no
need to have it if we use the new comment markers.

Rename ParseError to InvalidContestPage and avoid duplicating error messages if
we can find a 'botinfo' template on the contest page.