Fix starting of jobs from webinterface by switching back to gridengine
Since we cannot access jsub from Kubernetes, I'm switching back to
gridegine for now, but will stick with uwsgi as webserver. By using
uwsgi-plain, we can add our own options to uwsgi.ini, but I stick
to the standard convention where possible (location of virtualenv
and entrypoint).

Dropping Kubernetes also means we get Python 3.5, not Python 3.4. And
that we only need a single virtualenv, not two.